Updated to work with SFML master
Changed the setColor and getColor of sf::Text by setFillColor and getFillColor functions. The former are obsolete.
Added the functions corresponding to the outline color and outline thickness.
Added some overloads of operator << using the OutlineColor and OutlineThickness nested structs.
Fixed bug that did not use the correct kerning between texts within the same line.